move graphics presets into GothicGraphicsState.cpp
